refactor(core): extract diff workflow owner
