+++ ci/proof.toml
@@ -176,6 +176,7 @@ proof = [
   "cargo test -p tokmd-core lang_workflow --verbose",
   "cargo test -p tokmd-core module_workflow --verbose",
   "cargo test -p tokmd-core export_workflow --verbose",
+  "cargo test -p tokmd-core diff_workflow --verbose",
   "cargo test -p tokmd-core ffi",
   "cargo test -p tokmd-core --test ffi_parity_w53",
 ]
