+++ crates/tokmd-core/src/workflows/mod.rs
@@ -1,9 +1,11 @@
 //! Public workflow facade owner modules.
 
+mod diff;
 mod export;
 mod lang;
 mod module;
 
+pub use diff::diff_workflow;
 pub use export::{export_workflow, export_workflow_from_inputs};
 pub use lang::{lang_workflow, lang_workflow_from_inputs};
 pub use module::{module_workflow, module_workflow_from_inputs};
